refactor: organize runtime boundary modules
Move inspection structs into a canonical Squidie.Inspection namespace, group journal command implementations under Runtime.Journal.Commands, and extract public runtime option routing out of the Squidie facade. Keep the existing Runs and Runtime.Journal module names as documented compatibility wrappers so host callers retain their current API shape.
